refactor: add cluster table
