+++ server/common/common-dao/src/main/java/io/holoinsight/server/common/dao/converter/CustomConverter.java
@@ -1,11 +1,11 @@
 /*
  * Copyright 2022 Holoinsight Project Authors. Licensed under Apache-2.0.
  */
-package io.holoinsight.server.home.dal.converter;
+package io.holoinsight.server.common.dao.converter;
 
-import io.holoinsight.server.home.dal.model.Cluster;
-import io.holoinsight.server.home.dal.model.dto.ClusterDTO;
-import io.holoinsight.server.home.dal.transformer.BooleanIntegerMapper;
+import io.holoinsight.server.common.dao.entity.Cluster;
+import io.holoinsight.server.common.dao.entity.dto.ClusterDTO;
+import io.holoinsight.server.common.dao.transformer.BooleanIntegerMapper;
 import org.mapstruct.Mapper;
 
 import java.util.List;
