+++ server/common/common-dao/src/main/java/io/holoinsight/server/common/dao/entity/dto/ClusterDTO.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2022 Holoinsight Project Authors. Licensed under Apache-2.0.
  */
-package io.holoinsight.server.home.dal.model.dto;
+package io.holoinsight.server.common.dao.entity.dto;
 
 import lombok.Data;
