+++ server/common/common-service/src/main/java/io/holoinsight/server/common/model/CLUSTER_ROLE_CONST.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2022 Holoinsight Project Authors. Licensed under Apache-2.0.
  */
-package io.holoinsight.server.home.common.util;
+package io.holoinsight.server.common.model;
 
 /**
  *
