+++ server/home/home-alert/src/main/java/io/holoinsight/server/home/alert/service/task/coordinator/CoordinatorService.java
@@ -10,9 +10,9 @@ import io.holoinsight.server.common.J;
 import io.holoinsight.server.home.alert.service.task.CacheAlertTask;
 import io.holoinsight.server.home.alert.service.task.coordinator.server.NettyServer;
 import io.holoinsight.server.home.biz.common.MetaDictUtil;
-import io.holoinsight.server.home.common.util.CLUSTER_ROLE_CONST;
-import io.holoinsight.server.home.dal.mapper.ClusterMapper;
-import io.holoinsight.server.home.dal.model.Cluster;
+import io.holoinsight.server.common.model.CLUSTER_ROLE_CONST;
+import io.holoinsight.server.common.dao.mapper.ClusterMapper;
+import io.holoinsight.server.common.dao.entity.Cluster;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
