+++ server/home/home-task/src/main/java/io/holoinsight/server/home/task/MonitorTaskManager.java
@@ -8,13 +8,13 @@ import io.holoinsight.server.common.J;
 import io.holoinsight.server.common.config.ProdLog;
 import io.holoinsight.server.common.config.ScheduleLoadTask;
 import io.holoinsight.server.home.biz.common.MetaDictUtil;
-import io.holoinsight.server.home.biz.service.ClusterService;
+import io.holoinsight.server.common.service.ClusterService;
 import io.holoinsight.server.home.biz.service.ClusterTaskService;
 import io.holoinsight.server.home.biz.service.impl.ClusterTaskServiceImpl;
-import io.holoinsight.server.home.common.util.CLUSTER_ROLE_CONST;
+import io.holoinsight.server.common.model.CLUSTER_ROLE_CONST;
 import io.holoinsight.server.home.common.util.Debugger;
 import io.holoinsight.server.home.dal.model.ClusterTask;
-import io.holoinsight.server.home.dal.model.dto.ClusterDTO;
+import io.holoinsight.server.common.dao.entity.dto.ClusterDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
